Remove license
You are now no longer allowed to use this software.
This decision hurts a little bit but it does allow me some more leverage towards Ultimaker to make sure that they eventually use this software in a free manner.

Previous versions of the software can be used freely, but are pretty much useless for anyone.