Save file when print is modified
To also save it when one of these lists or dictionaries for the extruder settings or evaluation is modified, I would either need to import my autolistener from Luna or force the programmer to modify the evaluation through a function which is a bit of a pain. I think I'll just force the programmer to remember for now that they have to save the print after the evaluation is modified.