Make name notifyable
Just to shut the QML warning up really. It normally already gets a notification that the name changed because the print itself changed. We don't change the names of prints while it's firmly in the model, but we do change them when creating the instances so we can't make it constant either.