Show time and date in prints menu
Because quite often you'll print the same thing multiple times in order to compare their quality, and on the same printer. By default, these prints will get the same name from Cura. So you then need to disambiguate between these prints. We can do that by giving the date that it was printed on (or at least saved).