Use more human readable time and date format
This is now displayed in the interface, so we'll need a time/date format that is a bit easier to read than the previous 180813-005916. It now looks like 2018-08-13_00-59-16. Should be clear. It can't use the more common format with spaces and colons and slashes and all that, because this same key is used to save the file and colons and slashes can't be seen in the file name.