Split data parameter up into X and Y
These are the inputs of the formula for least squares: B = (X^T * X)^-1 * X^T * Y. X is the predictor, Y is the response, B is the resulting list of multipliers. Should be simple to implement now, since most of the work is getting the input data in the correct format.